Fixed excel error
Fixed an error in the excel input spreadsheet regarding the names of
columns. I tried to include units previously, but this threw off the
reading functionality. Now the units for a column are listed below the
last value of the column. Also progressed slightly on the reading
functionality, but it isn't quite done yet. No error messages, but a
logic error somewhere.