Removed CO2Loc and Added Distribution Flexibility
Removed all code pertaining to CO2 locations hopefully, as it is no
longer used in this version. Also added the framework for eventually
using non-normal distributions by adding a distributions input into the
input sheet and adding some if statements to handle it. It can only
handle normal right now, but rayleigh will be added shortly.